Move font management to main()
Functions in the `sdl2::ttf` module require an API context (`Sdl2TtfContext`)
that initializes/deinitializes the underlying C API.

For that reason, structs like `ttf::Font` require a lifetime annotation
correlated to such context.

Now, it is hard to handle "parent" and "child" data (in the lifetime
meaning) as fields of the struct--and if you think about it, that's not
the most logical thing to do, since larger lifetimes are supposed to
"encapsulate" shorter ones.

For that reason, the `Sdl2TtfContext` object was moved to `main()`, it
is created there and passed as a parameter to the `Game` constructor.
This way, the lifetimes are easily verified, since both the ttf context
and the `Game` object have the same scope (that of `main()`).

Also, the actual font loading and initialization is done in
`Game::new()`, and the font is saved as a field of the returned `Game`,
so that other methods just access the stored font via `self`.

If different font styles are needed at some point, all of them could be
initialized in the same way in `Game::new()` and added as fields (or as
a "font dictionary" field.

Other changes:

- Also, moved documentation comments on top of the enums/structs they
  document, so that the doc tools identify them properly.

- Fixed global typo loose -> lose
- Renamed function `rec_generator` to `create_rect`