Replace sleep by a callback timer
The timer callback pushes a user-defined event in the queue,
this way the play loop just waits for the next event of whatever type
and if its a timer event, it updates the game evolution.

The timer events can be consumed by loops in any state (currently
only the play loop uses it).

Also implemented the PAUSED state (a rudimentary version).